Add support for embedding disk image as resource
This adds a C define `DISK_IMAGE_AS_RESOURCE` that will make the
interpreter look for the disk image as embedded resource `game/sundog.st`,
instead of one passed as parameter.

Don't add build system support for this, as we don't want to strongly
couple the build process to the proprietary disk image. This would be an
optional thing when switching to a configurable build system (see #15).