Move mouse event handling out of game_screen, into main
This removes the reliance on questionable SDL cross-thread behavior,
and makes game_screen less dependent on SDL (I think it's only still
used for the cursor switching, and mutexes).

Mouse state is passed to the screen in the event handler, and protected
by a mutex as it should be. No more "spooky action at a distance".

Also moves input bypass flag for debug UI up a layer.