build: Set specific C++ standard to use (C++11)
There appears to be a difference in interpretation between clang and
gcc. Whereas gcc will by default pick the newest standard, clang picks
the oldest (or something like that). Enforce a specific standard.

It's good to be explicit but this is not really needed yet. It will be
when upgrading imgui, whose newer version uses constexpr and such.