Implement original disk behavior, remove copy protection workaround
Add a pre-access hook for disk access, and emulate disk controller
behavior as necessary to satisfy the game's copy protection. This makes
it possible to remove the periodic poking into gamestate to defuse it.

Thanks to Wayne Holder for explaining how the system works, it would
have been much harder to figure out without that.

This removes the last TODO.